+++ exonum-java-binding/qa-service/src/test/java/com/exonum/binding/qaservice/transactions/IncrementCounterTxIntegrationTest.java
@@ -23,10 +23,10 @@ import static com.exonum.binding.qaservice.transactions.ContextUtils.newContext;
 import static com.exonum.binding.qaservice.transactions.CreateCounterTxIntegrationTest.createCounter;
 import static com.exonum.binding.qaservice.transactions.IncrementCounterTx.converter;
 import static com.exonum.binding.qaservice.transactions.QaTransaction.INCREMENT_COUNTER;
+import static com.exonum.binding.qaservice.transactions.TransactionError.UNKNOWN_COUNTER;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.exonum.binding.common.hash.HashCode;
@@ -37,12 +37,12 @@ import com.exonum.binding.qaservice.QaService;
 import com.exonum.binding.storage.database.Database;
 import com.exonum.binding.storage.database.Fork;
 import com.exonum.binding.storage.database.MemoryDb;
-import com.exonum.binding.storage.indices.MapIndex;
 import com.exonum.binding.storage.indices.ProofMapIndexProxy;
 import com.exonum.binding.test.Bytes;
 import com.exonum.binding.test.RequiresNativeLibrary;
 import com.exonum.binding.transaction.RawTransaction;
 import com.exonum.binding.transaction.TransactionContext;
+import com.exonum.binding.transaction.TransactionExecutionException;
 import com.exonum.binding.util.LibraryLoader;
 import com.google.gson.reflect.TypeToken;
 import nl.jqno.equalsverifier.EqualsVerifier;
@@ -88,7 +88,7 @@ class IncrementCounterTxIntegrationTest {
 
   @Test
   @RequiresNativeLibrary
-  void executeIncrementsCounter() throws CloseFailuresException {
+  void executeIncrementsCounter() throws Exception {
     try (Database db = MemoryDb.newInstance();
         Cleaner cleaner = new Cleaner()) {
       Fork view = db.createFork(cleaner);
@@ -130,14 +130,9 @@ class IncrementCounterTxIntegrationTest {
       // Execute the transaction
       IncrementCounterTx tx = new IncrementCounterTx(seed, nameHash);
       TransactionContext context = newContext(view);
-      tx.execute(context);
-
-      // Check there isn’t such a counter after tx execution
-      QaSchema schema = new QaSchema(view);
-      MapIndex<HashCode, Long> counters = schema.counters();
-      MapIndex<HashCode, String> counterNames = schema.counterNames();
-      assertFalse(counters.containsKey(nameHash));
-      assertFalse(counterNames.containsKey(nameHash));
+      TransactionExecutionException e = assertThrows(TransactionExecutionException.class,
+          () -> tx.execute(context));
+      assertThat(e.getErrorCode(), equalTo(UNKNOWN_COUNTER.code));
     }
   }